+++ hs/app/BoolTypes.hs
@@ -1,0 +1,5 @@
+module BoolTypes where
+
+type PrintLogs     = Bool
+type Compile       = Bool
+type Exec          = Bool